Implement faceting search
This finds the single path to a 7/7/4 with the highest probability.
I'm not sure if this is the value we want to maximize: for example,
you could consider the path that maximizes total probability of all
children with 7/7/4 rolls, or the path that maximizes minimum number
of successful rolls.

I'm also not sure if the optimal decision changes as you reveal one
outcome at a time. I suspect it does, but have no evidence.