Implement brute-force analysis of game tree
The efficiency relies on the fact that order doesn't matter. For example,
if faceting the first line, the following two rolls are equivalent:

```
SUCCESS | FAILURE
FAILURE | SUCCESS
```

This allows us to memoize on a smaller set of states, using the
current success chance, successes per line, and rolls per line. The
maximum number of states should be about 66^3 * 6, where 66 is
`(1..11).sum()` for the number of success or failures per line per
total rolls, and 6 is the number of distinct chances (25, 35, ..., 75).