Remove big integer dependency
This removes most allocations and drops runtime down to about 0.3s.

In order to get all values to fit inside 128 bits, we reduce the
`Chance` fraction from `(10n + 25) / 100` to `(2n + 5) / 20`.
Some back-of-the-envelope math shows that the maximum numerator
should be about 15^30 ~ (2^4)^30 ~ 2^120, which just barely fits
inside a `u128`.

In order to remove the dependency on rational numbers, we note that
the denominator is constant (20^30) and identical for both success
and failure branches at each level of recursion, so we can just do
computations on the numerator and normalize at the end.

This did introduce a tricky bug for counting terminal states. Because
the denominator is now implicit, we can only return non-zero
expectations at leaf nodes, or else we'll miss extra factors of 20.