Fix napkin math
So the maximum numerator should actually be 20^30, and not 15^30,
meaning that 128 bits is not sufficient to hold all possible numerators.
This commit switches to the `uint` crate for fixed-width big integers,
which still avoids heap allocation.

Performance seems about the same as using `u128`, and using 192 bits
should be enough for up to 44 total rolls (there are only 30 max in
the game as of now).